requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
chromadb
beautifulsoup4
python-dotenv
tiktoken
langchain-text-splitters
sentence-transformers
lxml 
